+++ web/src/beta/features/Editor/tabs/publish/Nav/PublishModal/index.tsx
@@ -1,9 +1,9 @@
 import { useMemo } from "react";
 
 import Button from "@reearth/beta/components/Button";
+import ToggleButton from "@reearth/beta/components/fields/Toggle";
 import Icon from "@reearth/beta/components/Icon";
 import Modal from "@reearth/beta/components/Modal";
-import ToggleButton from "@reearth/beta/components/properties/Toggle";
 import Text from "@reearth/beta/components/Text";
 import { useT } from "@reearth/services/i18n";
 import { styled, metricsSizes, useTheme } from "@reearth/services/theme";
